doubles are differnt sig figs
The idea is that some numerical methods will guarantee you a number
of significant figures over a wide range of magnitudes. In that case,
the usual assertEquals with a tolerance doesn't really do the job
and you're stuck with a semantically odd use of assertTrue. This
will go to rectifying that.